Seperate HTTP into its own namespace.
