+++ .gitignore
@@ -5,3 +5,4 @@ tests/test
 build
 lxbuild
 winbuild
+.vscode
