+++ http-curl.cpp
@@ -7,10 +7,6 @@
 // found online at https://opensource.org/licenses/MIT.
 //
 
-#include "analytics.h"
-
-#include "json.hpp"
-
 #include "http-curl.hpp"
 
 #include <cstring>
@@ -21,89 +17,83 @@
 
 #include <curl/curl.h>
 
-// To keep this simple, we include the entire implementation in this one
-// C++ file.  Organizationally we'd probably have rather split this up,
-// but having a single file to integrate makes it easier to integrate
-// into other projects.
-
-using json = nlohmann::json;
-
 namespace segment {
+namespace http {
 
-// Internal curl exception error.  Most of these will be ENOMEM.
-class curlExcept : public std::exception {
-public:
-    curlExcept(CURLcode rv)
-    {
-        this->code = rv;
+    // Internal curl exception error.  Most of these will be ENOMEM.
+    class curlExcept : public std::exception {
+    public:
+        curlExcept(CURLcode rv)
+        {
+            this->code = rv;
+        };
+        const char* what() const throw()
+        {
+            return (curl_easy_strerror(this->code));
+        }
+        CURLcode code;
     };
-    const char* what() const throw()
-    {
-        return (curl_easy_strerror(this->code));
-    }
-    CURLcode code;
-};
 
-class curlReq {
-public:
-    curlReq()
-    {
-        this->headers = NULL;
-        this->req = NULL;
-    }
-    ~curlReq()
-    {
-        if (this->headers != NULL) {
-            curl_slist_free_all(this->headers);
+    class curlReq {
+    public:
+        curlReq()
+        {
+            this->headers = NULL;
+            this->req = NULL;
         }
-        if (this->req != NULL) {
-            curl_easy_cleanup(this->req);
-        }
-    }
-
-    void addHeader(const std::string key, const std::string value)
-    {
-        struct curl_slist* hdrs;
-        std::string line;
-
-        line = key + ": " + value;
-        hdrs = curl_slist_append(this->headers, line.c_str());
-        if (hdrs == NULL) {
-            throw std::bad_alloc(); // cannot think of any other reason...
-        }
-        this->headers = hdrs;
-    }
-
-    void setBody(const std::string& body)
-    {
-        this->body = body;
-    }
-
-    static size_t writeCallback(char* ptr, size_t sz, size_t nmemb, void* udata)
-    {
-        curlReq* creq = (curlReq*)udata;
-        char* ucdata = (char*)udata;
-        size_t nbytes = sz * nmemb;
-        creq->respData += std::string(ucdata, nbytes);
-        return (nbytes);
-    }
-
-    void perform(const std::string meth, const std::string url)
-    {
-        CURL* req;
-        CURLcode rv;
-        long code;
-
-        if (meth != "POST") {
-            throw std::invalid_argument("Only POST supported");
+        ~curlReq()
+        {
+            if (this->headers != NULL) {
+                curl_slist_free_all(this->headers);
+            }
+            if (this->req != NULL) {
+                curl_easy_cleanup(this->req);
+            }
         }
 
-        if (this->req == NULL) {
-            this->req = curl_easy_init();
+        void addHeader(const std::string key, const std::string value)
+        {
+            struct curl_slist* hdrs;
+            std::string line;
+
+            line = key + ": " + value;
+            hdrs = curl_slist_append(this->headers, line.c_str());
+            if (hdrs == NULL) {
+                throw std::bad_alloc(); // cannot think of any other reason...
+            }
+            this->headers = hdrs;
         }
-        if ((req = this->req) == NULL) {
-            throw std::bad_alloc();
+
+        void setBody(const std::string& body)
+        {
+            this->body = body;
         }
+
+        static size_t writeCallback(char* ptr, size_t sz, size_t nmemb, void* udata)
+        {
+            curlReq* creq = (curlReq*)udata;
+            char* ucdata = (char*)udata;
+            size_t nbytes = sz * nmemb;
+            creq->respData += std::string(ucdata, nbytes);
+            return (nbytes);
+        }
+
+        void perform(const std::string meth, const std::string url)
+        {
+            CURL* req;
+            CURLcode rv;
+            long code;
+
+            if (meth != "POST") {
+                throw std::invalid_argument("Only POST supported");
+            }
+
+            if (this->req == NULL) {
+                this->req = curl_easy_init();
+            }
+            if ((req = this->req) == NULL) {
+                throw std::bad_alloc();
+            }
 
 #define setopt(k, v)                                        \
     if ((rv = curl_easy_setopt(req, k, v)) != CURLE_OK) {   \
@@ -114,23 +104,23 @@ public:
         }                                                   \
     }
 
-        std::string body(this->body.begin(), this->body.end());
+            std::string body(this->body.begin(), this->body.end());
 
-        // We only handle post.
-        setopt(CURLOPT_URL, url.c_str());
-        setopt(CURLOPT_POSTFIELDS, body.c_str());
-        setopt(CURLOPT_POSTFIELDSIZE, body.length());
-        setopt(CURLOPT_POST, 1);
-        setopt(CURLOPT_HTTPHEADER, this->headers);
-        setopt(CURLOPT_WRITEFUNCTION, this->writeCallback);
-        setopt(CURLOPT_WRITEDATA, this);
-        //setopt(CURLOPT_VERBOSE, 1);
+            // We only handle post.
+            setopt(CURLOPT_URL, url.c_str());
+            setopt(CURLOPT_POSTFIELDS, body.c_str());
+            setopt(CURLOPT_POSTFIELDSIZE, body.length());
+            setopt(CURLOPT_POST, 1);
+            setopt(CURLOPT_HTTPHEADER, this->headers);
+            setopt(CURLOPT_WRITEFUNCTION, this->writeCallback);
+            setopt(CURLOPT_WRITEDATA, this);
+            //setopt(CURLOPT_VERBOSE, 1);
 
-        if ((rv = curl_easy_perform(req)) != CURLE_OK) {
-            if (rv == CURLE_OUT_OF_MEMORY) {
-                throw std::bad_alloc();
+            if ((rv = curl_easy_perform(req)) != CURLE_OK) {
+                if (rv == CURLE_OUT_OF_MEMORY) {
+                    throw std::bad_alloc();
+                }
             }
-        }
 
 #define getinfo(k, v)                                      \
     if ((rv = curl_easy_getinfo(req, k, v)) != CURLE_OK) { \
@@ -139,51 +129,52 @@ public:
         }                                                  \
     }
 
-        // get status
-        getinfo(CURLINFO_RESPONSE_CODE, &code);
-        if (code == 0) {
-            long errn;
-            getinfo(CURLINFO_OS_ERRNO, &errn);
-            throw std::system_error((int)errn, std::system_category());
+            // get status
+            getinfo(CURLINFO_RESPONSE_CODE, &code);
+            if (code == 0) {
+                long errn;
+                getinfo(CURLINFO_OS_ERRNO, &errn);
+                throw std::system_error((int)errn, std::system_category());
+            }
+            if (code > 299) {
+                throw HttpError(code);
+            }
+            this->respCode = (int)code;
         }
-        if (code > 299) {
-            throw HttpError(code);
+
+        // This is probably not terribly idiomatic C++, but we put them
+        // here to make things easy.  C++ class friends would be better,
+        // but we're leaving this here in the implementation, so it should
+        // be fine.  These are used by the HttpHandlerCurl class.
+
+        std::string respData;
+        std::string respMessage;
+        int respCode;
+
+    private:
+        struct curl_slist* headers;
+        std::string body;
+        CURL* req;
+    };
+
+    std::shared_ptr<HttpResponse> HttpHandlerCurl::Handle(const HttpRequest& req)
+    {
+
+        auto resp = std::make_shared<HttpResponse>();
+        curlReq creq;
+
+        for (auto const& item : req.Headers) {
+            creq.addHeader(item.first, item.second);
         }
-        this->respCode = (int)code;
+
+        creq.setBody(req.Body);
+        creq.perform("POST", req.URL);
+        resp->Code = creq.respCode;
+        resp->Message = creq.respMessage;
+        resp->Body = creq.respData;
+
+        return resp;
     }
 
-    // This is probably not terribly idiomatic C++, but we put them
-    // here to make things easy.  C++ class friends would be better,
-    // but we're leaving this here in the implementation, so it should
-    // be fine.  These are used by the HttpHandlerCurl class.
-
-    std::string respData;
-    std::string respMessage;
-    int respCode;
-
-private:
-    struct curl_slist* headers;
-    std::string body;
-    CURL* req;
-};
-
-std::shared_ptr<HttpResponse> HttpHandlerCurl::Handle(const HttpRequest& req)
-{
-
-    auto resp = std::make_shared<HttpResponse>();
-    curlReq creq;
-
-    for (auto const& item : req.Headers) {
-        creq.addHeader(item.first, item.second);
-    }
-
-    creq.setBody(req.Body);
-    creq.perform("POST", req.URL);
-    resp->Code = creq.respCode;
-    resp->Message = creq.respMessage;
-    resp->Body = creq.respData;
-
-    return resp;
-}
-
+} // namespace http
 } // namespace segment
